Create METClassic project to separate out game specific files
