+++ src/MEU.GV4.Data/Serialization/ErrorConstants.cs
@@ -1,5 +1,5 @@
-﻿namespace MEU.GV4.Data.Providers;
-internal struct ErrorConstants
+﻿namespace MEU.GV4.Data.Serialization;
+public struct ErrorConstants
 {
     public const string FILE_CONTENTS_EMPTY = "File contents are empty";
     public const string INVALID_GRAPEVINE_FILE = "Invalid grapevine file";
