+++ src/MEU.GV4.Data/Serialization/GrapevineProviderException.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace MEU.GV4.Data.Providers;
+﻿namespace MEU.GV4.Data.Serialization;
 
 public class GrapevineProviderException : Exception
 {
