+++ src/MEU.GV4.METClassic/MEU.GV4.METClassic.csproj
@@ -1,0 +1,13 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net9.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\MEU.GV4.Data\MEU.GV4.Data.csproj" />
+  </ItemGroup>
+
+</Project>
