+++ src/MEU.GV4.METClassic/Models/Vampire.cs
@@ -1,4 +1,6 @@
-﻿namespace MEU.GV4.Data.Models.METClassic;
+﻿using MEU.GV4.Data.Models;
+
+namespace MEU.GV4.METClassic.Models;
 
 public class Vampire : METCharacter
 {
