+++ tests/MEU.GV4.Data.Tests/Providers/GrapevineLegacyXMLReaderCharacterTests.cs
@@ -1,6 +1,6 @@
 ﻿using MEU.GV4.Data.Models;
-using MEU.GV4.Data.Models.METClassic;
-using MEU.GV4.Data.Providers;
+using MEU.GV4.METClassic.Models;
+using MEU.GV4.METClassic.Serialization;
 using System.Xml.Linq;
 
 namespace MEU.GV4.Data.Tests.Providers;
