+++ tests/MEU.GV4.Data.Tests/Providers/GrapevineLegacyXMLReaderCommonTraitsTests.cs
@@ -1,5 +1,5 @@
 ﻿using MEU.GV4.Data.Models;
-using MEU.GV4.Data.Providers;
+using MEU.GV4.METClassic.Serialization;
 using System.Xml.Linq;
 
 namespace MEU.GV4.Data.Tests.Providers;
